Port libcamera changes for renaming sdl_texture_yuyv.{cpp,h} to sdl_texture_yuv.{cpp,h}.
